changed id from int to string

+++ packages/node/src/modules/server/broadcast.v
@@ -50,7 +50,7 @@ pub fn (mut app App) broadcast_route() vweb.Result {
 			println("[Broadcaster] Have not seen message before.\n[Broadcaster] Saving message to database.")
 			
 			message_db := database.Message_Table{
-				id: rand.uuid_v4().int()
+				id: rand.uuid_v4()
 				timestamp: decoded.message.time
 				contents: decoded.message.data
 				sender: parsed_sender
@@ -139,7 +139,7 @@ pub fn send_message(db database.DatabaseConnection, data string, receiver []u8) 
 
 	println("[Database] Saving message to database.")
 	db_msg := database.Message_Table{
-		id: rand.uuid_v4().int()
+		id: rand.uuid_v4()
 		timestamp: contents.time
 		sender: contents.sender.str()
 		receiver: contents.receiver.str()
